fixed more warnings, re-decomped a couple of functions, begin D231

+++ decompile/General/230/230_56_MM_TrackSelect_MenuProc.c
@@ -543,7 +543,7 @@ void DECOMP_MM_TrackSelect_MenuProc(struct RectMenu* menu)
 		{
 			iVar10++;
 
-			if (uVar15 <= iVar10)
+			if ((int)uVar15 <= iVar10)
 			{
 				iVar10 = 0;
 			}
